Moved SMC conditions to separate file for clarity
