+++ include/PQL/Expressions.h
@@ -271,70 +271,6 @@ namespace unfoldtacpn {
             void visit(Visitor&) const override;
         };
 
-        struct SMCSettings {
-            enum SMCBoundType { TimeBound, StepBound };
-            SMCBoundType boundType;
-            int bound;
-            float falsePositives;
-            float falseNegatives;
-            float indifferenceRegionUp;
-            float indifferenceRegionDown;
-            float defaultRate;
-            float confidence;
-            float estimationIntervalWidth;
-        };
-
-        class ProbaCondition : public SimpleQuantifierCondition {
-        public:
-            ProbaCondition(SMCSettings settings, Condition_ptr cond)
-            : SimpleQuantifierCondition(cond) {
-                _settings = settings;
-            }
-            void analyze(NamingContext& context) override;
-            virtual const SMCSettings settings() const { return _settings; }
-        protected:
-            SMCSettings _settings;
-        };
-
-        class PFCondition : public ProbaCondition {
-        public:
-            using ProbaCondition::ProbaCondition;
-            void visit(Visitor&) const override;
-        };
-
-        class PGCondition : public ProbaCondition {
-        public:
-            using ProbaCondition::ProbaCondition;
-            void visit(Visitor&) const override;
-        };
-
-        /*class StaticProbaCompCondition : public SimpleQuantifierCondition {
-        public:
-            StaticProbaCompCondition(Condition_ptr cond, double comp)
-            : SimpleQuantifierCondition(cond) {
-                _comp = comp;
-            }
-            void visit(Visitor&) const override;
-            virtual const double bound() const { return _comp; }
-        protected:
-            double _comp;
-        };
-
-        class ProbaCompCondition : public QuantifierCondition {
-        public:
-            ProbaCompCondition(Condition_ptr cond1, Condition_ptr cond2) {
-                _cond1 = cond1; 
-                _cond2 = cond2;
-            }
-            void visit(Visitor&) const override;
-            void analyze(NamingContext& context) override;
-            virtual const Condition_ptr& operator[] (size_t i) const override
-            { if(i == 0) return _cond1; return _cond2;}
-        protected:
-            Condition_ptr _cond1;
-            Condition_ptr _cond2;
-        };*/
-
         /******************** CONDITIONS ********************/
         class LogicalCondition : public Condition {
         public:
