+++ src/PQL/Expressions.cpp
@@ -98,11 +98,6 @@ namespace unfoldtacpn {
             _cond2->analyze(context);
         }
 
-        void ProbaCondition::analyze(NamingContext &context)
-        {
-            _cond->analyze(context);
-        }
-
         /*void ProbaCompCondition::analyze(NamingContext &context) 
         {
             _cond1->analyze(context);
@@ -177,16 +172,6 @@ namespace unfoldtacpn {
         }
 
         void AGCondition::visit(Visitor& ctx) const
-        {
-            ctx.accept<decltype(this)>(this);
-        }
-
-        void PFCondition::visit(Visitor& ctx) const
-        {
-            ctx.accept<decltype(this)>(this);
-        }
-
-        void PGCondition::visit(Visitor& ctx) const
         {
             ctx.accept<decltype(this)>(this);
         }
